+++ Mile.Helpers/Mile.Helpers.CppBase.h
@@ -49,6 +49,17 @@ namespace Mile
     std::string VFormatString(
         _In_z_ _Printf_format_string_ char const* const Format,
         _In_z_ _Printf_format_string_ va_list ArgList);
+
+    /**
+     * @brief Write formatted data to a onebyte or multibyte string, suggested
+     *        encoding with UTF-8.
+     * @param Format Format-control string.
+     * @param ... Optional arguments to be formatted.
+     * @return A formatted string if successful, an empty string otherwise.
+    */
+    std::string FormatString(
+        _In_z_ _Printf_format_string_ char const* const Format,
+        ...);
 }
 
 #endif // !MILE_WINDOWS_HELPERS_CPPBASE