+++ ReleaseNotes.md
@@ -34,3 +34,4 @@
 - Add Mile::VFormatWideString method.
 - Add Mile::FormatWideString method.
 - Add Mile::VFormatString method.
+- Add Mile::FormatString method.